+++ grails-app/controllers/io/xh/hoist/impl/XhController.groovy
@@ -19,6 +19,7 @@ import io.xh.hoist.pref.PrefService
 import io.xh.hoist.pref.Preference
 import io.xh.hoist.security.AccessAll
 import io.xh.hoist.security.BaseAuthenticationService
+import io.xh.hoist.telemetry.TraceService
 import io.xh.hoist.track.TrackService
 import io.xh.hoist.environment.EnvironmentService
 import io.xh.hoist.user.BaseUserService
@@ -40,6 +41,7 @@ class XhController extends BaseController {
     EnvironmentService environmentService
     BaseUserService userService
     ClusterService clusterService
+    TraceService traceService
 
 
     //------------------------
@@ -157,6 +159,16 @@ class XhController extends BaseController {
             prefService.clearPreferences()
             viewService.clearAllState()
         }
+        renderSuccess()
+    }
+
+    //------------------------
+    // Traces
+    //------------------------
+    def submitSpans() {
+        ensureClientUsernameMatchesSession()
+        def payload = parseRequestJSONArray()
+        traceService.submitClientSpans(payload as List<Map>)
         renderSuccess()
     }