+++ src/main/groovy/io/xh/hoist/HoistFilter.groovy
@@ -16,6 +16,7 @@ import jakarta.servlet.http.HttpServletRequest
 import jakarta.servlet.http.HttpServletResponse
 
 import static io.xh.hoist.util.Utils.authenticationService
+import static io.xh.hoist.util.Utils.traceService
 import static io.xh.hoist.util.Utils.getClusterService
 
 /**
@@ -35,7 +36,7 @@ class HoistFilter implements Filter, LogSupport {
         HttpServletRequest httpRequest = (HttpServletRequest) request
         HttpServletResponse httpResponse = (HttpServletResponse) response
 
-        try {
+        try (def scope = traceService.restoreContextFromRequest(httpRequest)) {
             clusterService.ensureRunning()
             if (authenticationService.allowRequest(httpRequest, httpResponse)) {
                 chain.doFilter(request, response)